fix: amending all not working functions to adapt MCV architecture
