+++ index.html
@@ -139,7 +139,7 @@
         </div>
     </div>
 
-
+//-->
     <footer class="footer"></footer>
