+++ src/js/view.js
@@ -1,0 +1,18 @@
+let id = Math.random() + ''
+
+export const searchResultsContainer = document.querySelector('.search-results__container')
+export const mainContainerResults = document.querySelector('.search-results')
+
+export const renderResult = function (result) {
+	let html = ` <div class="search-results__result" data-id="${id}">
+<div class="search-results__result--icon"><i class="fa-solid fa-plus" style="color: #418900;"></i></div>
+<div class="search-results__heading">
+    <img src="${result.photo}" alt="Photo of the bird">
+    <p class="search-results__name">${result.name}</p>
+</div>
+</div>`
+
+	mainContainerResults.style.display = 'flex'
+	searchResultsContainer.insertAdjacentHTML('afterbegin', html)
+}
+
